+++ MoPubSDK/Native Ads/Internal/MPMoPubNativeAdAdapter.m
@@ -13,6 +13,9 @@
 #define kDefaultActionURLKey        @"clk"
 #define kClickTrackerURLKey         @"clktracker"
 
+static NSString *kDAAIconImageName = @"MPDAAIcon";
+static NSString *kDAAIconTapDestinationURL = @"https://www.mopub.com/optout";
+
 @interface MPMoPubNativeAdAdapter () <MPAdDestinationDisplayAgentDelegate>
 
 @property (nonatomic, readonly, strong) MPAdDestinationDisplayAgent *destinationDisplayAgent;
@@ -88,17 +91,33 @@
         return;
     }
 
-    self.rootViewController = controller;
     self.actionCompletionBlock = completionBlock;
 
     [self.destinationDisplayAgent displayDestinationForURL:URL];
+}
+
+#pragma mark - DAA Icon
+
+- (void)daaIconTapped
+{
+    [self.destinationDisplayAgent displayDestinationForURL:[NSURL URLWithString:kDAAIconTapDestinationURL]];
+}
+
+- (void)loadDAAIconIntoImageView:(UIImageView *)imageView
+{
+    imageView.image = [UIImage imageNamed:MPResourcePathForResource(kDAAIconImageName)];
+
+    // Attach a gesture recognizer to handle loading the daa icon URL.
+    UITapGestureRecognizer *tapRecognizer = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(daaIconTapped)];
+    imageView.userInteractionEnabled = YES;
+    [imageView addGestureRecognizer:tapRecognizer];
 }
 
 #pragma mark - <MPAdDestinationDisplayAgent>
 
 - (UIViewController *)viewControllerForPresentingModalView
 {
-    return self.rootViewController;
+    return [self.delegate viewControllerForPresentingModalView];
 }
 
 - (void)displayAgentWillPresentModal