+++ MoPubSampleApp/Domain/MPAdInfo.m
@@ -49,6 +49,16 @@
              ];
 }
 
++ (NSArray *)rewardedVideoAds
+{
+    return @[
+             [MPAdInfo infoWithTitle:@"Ad Colony" ID:@"16b039dc8c994ac89f0efe0a1bf77aa3" type:MPAdInfoRewardedVideo],
+             [MPAdInfo infoWithTitle:@"Chartboost" ID:@"8be0bb08fb4f4e90a86416c29c235d4a" type:MPAdInfoRewardedVideo],
+             [MPAdInfo infoWithTitle:@"Vungle" ID:@"48274e80f11b496bb3532c4f59f28d12" type:MPAdInfoRewardedVideo],
+             [MPAdInfo infoWithTitle:@"Unity" ID:@"091fb0bc95254e74a3f85753f589fdb7" type:MPAdInfoRewardedVideo]
+             ];
+}
+
 + (NSArray *)nativeAds
 {
     return @[