+++ MoPubSampleApp/Views/MPNativeAdCell.h
@@ -14,6 +14,7 @@
 @property (strong, nonatomic) UILabel *mainTextLabel;
 @property (strong, nonatomic) UIImageView *iconImageView;
 @property (strong, nonatomic) UIImageView *mainImageView;
+@property (strong, nonatomic) UIImageView *DAAIconImageView;
 @property (strong, nonatomic) UILabel *ctaLabel;
 
 @end