+++ MoPubSampleApp/Views/MPNativeAdView.h
@@ -13,6 +13,7 @@
 @property (weak, nonatomic) IBOutlet UILabel *ctaLabel;
 @property (weak, nonatomic) IBOutlet UIImageView *iconImageView;
 @property (weak, nonatomic) IBOutlet UIImageView *fullsizeImageView;
+@property (weak, nonatomic) IBOutlet UIImageView *DAAIconImageView;
 @property (weak, nonatomic) IBOutlet UILabel *titleLabel;
 @property (weak, nonatomic) IBOutlet UILabel *bodyLabel;
 @property (weak, nonatomic) IBOutlet UITapGestureRecognizer *tapGesture;