+++ Specs/MoPubSDKSpecs/ThirdPartyIntegrations/Vungle/VungleRewardedVideoCustomEventSpec.mm
@@ -223,7 +223,6 @@ describe(@"VungleRewardedVideoCustomEvent", ^{
             });
         });
     });
-
 });
 
 SPEC_END